+++ src/Magento/MagentoCloud/Config/DeploymentConfig.php
@@ -5,8 +5,45 @@
  */
 namespace Magento\MagentoCloud\Config;
 
-class Application
+use Magento\MagentoCloud\DB\Adapter;
+use Magento\MagentoCloud\Environment;
+use Magento\MagentoCloud\Shell\ShellInterface;
+use Psr\Log\LoggerInterface;
+
+class DeploymentConfig
 {
+    /**
+     * @var Environment
+     */
+    private $environment;
+
+    /**
+     * @var LoggerInterface
+     */
+    private $logger;
+
+    /**
+     * @var ShellInterface
+     */
+    private $shell;
+
+    /**
+     * @var Adapter
+     */
+    private $adapter;
+
+    public function __construct(
+        Environment $environment,
+        LoggerInterface $logger,
+        ShellInterface $shell,
+        Adapter $adapter
+    ) {
+        $this->environment = $environment;
+        $this->logger = $logger;
+        $this->shell = $shell;
+        $this->adapter = $adapter;
+    }
+
     /**
      * Verifies is Magento installed based on install date in env.php
      *
@@ -21,16 +58,16 @@ class Application
         //2. check if core_config_data and setup_module tables exist
         //3. check install date
 
-        $this->env->log('Checking if db exists and has tables');
-        $output = $this->executeDbQuery('SHOW TABLES');
+        $this->logger->info('Checking if db exists and has tables');
+        $output = $this->adapter->execute('SHOW TABLES');
         if (is_array($output) && count($output) > 1) {
             if (!in_array('core_config_data', $output) || !in_array('setup_module', $output)) {
-                $this->env->log('Missing either core_config_data or setup_module table');
+                $this->logger->info('Missing either core_config_data or setup_module table');
                 exit(5);
             } elseif (file_exists($configFile)) {
                 $data = include $configFile;
                 if (isset($data['install']) && isset($data['install']['date'])) {
-                    $this->env->log("Magento was installed on " . $data['install']['date']);
+                    $this->logger->info("Magento was installed on " . $data['install']['date']);
                     $isInstalled = true;
                 } else {
                     $config['install']['date'] = date('r');
@@ -39,7 +76,7 @@ class Application
                     $isInstalled = true;
                 }
             } else {
-                $this->env->execute('touch ' . $configFile);
+                $this->shell->execute('touch ' . $configFile);
                 $config['install']['date'] = date('r');
                 $updatedConfig = '<?php' . "\n" . 'return ' . var_export($config, true) . ';';
                 file_put_contents($configFile, $updatedConfig);
